requirements.txt
Flask
python-dotenv
requests
Flask-Cors